:recycle:(refactor): Change to move the types folder out of the components folder
