+++ src/api/ChartDataApi.ts
@@ -1,4 +1,4 @@
-import type { ChartDate } from "@/components/types/ChartDateaResponse";
+import type { ChartDate } from "@/types/ChartDateaResponse";
 import axios from "axios";
 
 const BASE_URL = 'http://localhost:8080/api/metrics/chart';
