+++ src/components/chartDataFilter/ChartDataFilter.vue
@@ -71,9 +71,9 @@ import './ChartDataFilter.css'
 import { onMounted, type Ref, ref } from 'vue'
 import Select from 'primevue/select'
 import DatePicker from 'primevue/datepicker'
-import type { FilterOptions } from '@/components/types/FilterOptions.ts'
+import type { FilterOptions } from '@/types/FilterOptions'
 import { fetchFilterOptions } from '@/api/FiltersApi.ts'
-import type { SelectListOption } from '@/components/types/SelectListOption.ts'
+import type { SelectListOption } from '@/types/SelectListOption'
 
 const baseDate = new Date()
