--- src/components/types/PaginatedInterfaces/ApiResponse.ts
@@ -1,8 +1,0 @@
-import type { PaginatedResponse } from '@/components/types/PaginatedInterfaces/PaginatedResponse.ts'
-import type { Product } from '@/components/types/Product.ts'
-import type { Company } from '@/components/types/Company.ts'
-
-export interface ApiResponse {
-  products: PaginatedResponse<Product>;
-  companies: PaginatedResponse<Company>;
-}
