+++ src/types/ChartDateaResponse.ts
@@ -2,4 +2,5 @@ export interface ChartDate {
   recidivismRate: number;
   ticketClosureTimeInHours: number;
   ticketsCount:number;
+  slaCompliancePercentualDto:number;
 }
